adding openapi docs
